+++ test/integration/test_easy_ship.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+require 'integration_helper'
+require 'mws/easy_ship'
+
+class TestEasyShip < IntegrationTest
+  def test_getting_service_status
+    clients.each do |client|
+      res = client.get_service_status
+      refute_empty res.parse
+    end
+  end
+end